requirements.txt
Flask == 3.0.3
python-dotenv == 1.0.1
psycopg2 == 2.9.9
Flask-WTF == 1.2.1
pandas  == 2.2.2
